requirements.txt
cffi==1.17.1
pycparser==2.22
PyNaCl==1.5.0
